Refactor dataset display to reduce file sizes.
